Convert more files
