+++ e2e/specs/confirmations/approve-custom-erc20.spec.ts
@@ -7,15 +7,13 @@ import {
   withFixtures,
   defaultGanacheOptions,
 } from '../../fixtures/fixture-helper';
-import {
-  SMART_CONTRACTS,
-  contractConfiguration,
-} from '../../../app/util/test/smart-contracts';
-import { ActivitiesViewSelectorsText } from '../../selectors/Transactions/ActivitiesView.selectors';
 
 import TabBarComponent from '../../pages/wallet/TabBarComponent';
 import TestDApp from '../../pages/Browser/TestDApp';
+import { SMART_CONTRACTS } from '../../../app/util/test/smart-contracts';
+import ContractApprovalBottomSheet from '../../pages/Browser/ContractApprovalBottomSheet';
 import Assertions from '../../utils/Assertions';
+import { ActivitiesViewSelectorsText } from '../../selectors/Transactions/ActivitiesView.selectors';
 import { mockEvents } from '../../api-mocking/mock-config/mock-events';
 import { buildPermissions } from '../../fixtures/utils';
 
@@ -24,14 +22,15 @@ const HST_CONTRACT = SMART_CONTRACTS.HST;
 describe(SmokeConfirmations('ERC20 tokens'), () => {
   beforeAll(async () => {
     jest.setTimeout(170000);
-    await TestHelpers.reverseServerPort();
+    if (device.getPlatform() === 'android') {
+      await TestHelpers.reverseServerPort();
+    }
   });
 
-  it('send an ERC20 token from a dapp', async () => {
+  it('approve custom ERC20 token amount from a dapp', async () => {
     const testSpecificMock  = {
       GET: [
-        mockEvents.GET.suggestedGasFeesApiGanache,
-        mockEvents.GET.remoteFeatureFlagsOldConfirmations
+        mockEvents.GET.suggestedGasFeesApiGanache
       ],
     };
 
@@ -47,34 +46,44 @@ describe(SmokeConfirmations('ERC20 tokens'), () => {
         smartContract: HST_CONTRACT,
         testSpecificMock,
       },
-      async ({ contractRegistry }) => {
+      // Remove any once withFixtures is typed
+      // eslint-disable-next-line @typescript-eslint/no-explicit-any
+      async ({ contractRegistry }: { contractRegistry: any }) => {
         const hstAddress = await contractRegistry.getContractAddress(
           HST_CONTRACT,
         );
         await loginToApp();
-
         // Navigate to the browser screen
         await TabBarComponent.tapBrowser();
+
         await TestDApp.navigateToTestDappWithContract({
           contractAddress: hstAddress,
         });
-        await TestHelpers.delay(3000);
+        await TestDApp.tapApproveERC20TokensButton();
 
-        // Transfer ERC20 tokens
-        await TestDApp.tapERC20TransferButton();
-        await TestHelpers.delay(3000);
+        //Input custom token amount
+        await Assertions.checkIfVisible(
+          ContractApprovalBottomSheet.approveTokenAmount,
+        );
+        await ContractApprovalBottomSheet.clearInput();
+        await ContractApprovalBottomSheet.inputCustomAmount('2');
 
-        // Tap confirm button
-        await TestDApp.tapConfirmButton();
+        // Assert that custom token amount is shown
+        await Assertions.checkIfElementToHaveText(
+          ContractApprovalBottomSheet.approveTokenAmount,
+          '2',
+        );
+        // Tap next button
+        await ContractApprovalBottomSheet.tapNextButton();
+
+        // Tap approve button
+        await ContractApprovalBottomSheet.tapApproveButton();
 
         // Navigate to the activity screen
         await TabBarComponent.tapActivity();
-
-        // Assert "Sent Tokens" transaction is displayed
+        // Assert erc20 is approved
         await Assertions.checkIfTextIsDisplayed(
-          ActivitiesViewSelectorsText.SENT_TOKENS_MESSAGE_TEXT(
-            contractConfiguration[HST_CONTRACT].tokenName,
-          ),
+          ActivitiesViewSelectorsText.CONFIRM_TEXT,
         );
       },
     );
