+++ e2e/specs/confirmations/approve-erc721.spec.ts
@@ -1,33 +1,31 @@
 'use strict';
+
 import { SmokeConfirmations } from '../../tags';
 import TestHelpers from '../../helpers';
 import { loginToApp } from '../../viewHelper';
+
+import TabBarComponent from '../../pages/wallet/TabBarComponent';
+import TestDApp from '../../pages/Browser/TestDApp';
 import FixtureBuilder from '../../fixtures/fixture-builder';
 import {
   withFixtures,
   defaultGanacheOptions,
 } from '../../fixtures/fixture-helper';
-
-import TabBarComponent from '../../pages/wallet/TabBarComponent';
-import TestDApp from '../../pages/Browser/TestDApp';
 import { SMART_CONTRACTS } from '../../../app/util/test/smart-contracts';
-import ContractApprovalBottomSheet from '../../pages/Browser/ContractApprovalBottomSheet';
-import Assertions from '../../utils/Assertions';
 import { ActivitiesViewSelectorsText } from '../../selectors/Transactions/ActivitiesView.selectors';
+import Assertions from '../../utils/Assertions';
 import { mockEvents } from '../../api-mocking/mock-config/mock-events';
 import { buildPermissions } from '../../fixtures/utils';
 
-const HST_CONTRACT = SMART_CONTRACTS.HST;
+describe(SmokeConfirmations('ERC721 tokens'), () => {
+  const NFT_CONTRACT = SMART_CONTRACTS.NFTS;
 
-describe(SmokeConfirmations('ERC20 tokens'), () => {
   beforeAll(async () => {
-    jest.setTimeout(170000);
-    if (device.getPlatform() === 'android') {
-      await TestHelpers.reverseServerPort();
-    }
+    jest.setTimeout(150000);
+    await TestHelpers.reverseServerPort();
   });
 
-  it('approve custom ERC20 token amount from a dapp', async () => {
+  it('approve an ERC721 token from a dapp', async () => {
     const testSpecificMock  = {
       GET: [
         mockEvents.GET.suggestedGasFeesApiGanache
@@ -43,43 +41,28 @@ describe(SmokeConfirmations('ERC20 tokens'), () => {
           .build(),
         restartDevice: true,
         ganacheOptions: defaultGanacheOptions,
-        smartContract: HST_CONTRACT,
+        smartContract: NFT_CONTRACT,
         testSpecificMock,
       },
-      async ({ contractRegistry }) => {
-        const hstAddress = await contractRegistry.getContractAddress(
-          HST_CONTRACT,
+      // Remove any once withFixtures is typed
+      // eslint-disable-next-line @typescript-eslint/no-explicit-any
+      async ({ contractRegistry }: { contractRegistry: any }) => {
+        const nftsAddress = await contractRegistry.getContractAddress(
+          NFT_CONTRACT,
         );
         await loginToApp();
         // Navigate to the browser screen
         await TabBarComponent.tapBrowser();
-
         await TestDApp.navigateToTestDappWithContract({
-          contractAddress: hstAddress,
+          contractAddress: nftsAddress,
         });
-        await TestDApp.tapApproveERC20TokensButton();
-
-        //Input custom token amount
-        await Assertions.checkIfVisible(
-          ContractApprovalBottomSheet.approveTokenAmount,
-        );
-        await ContractApprovalBottomSheet.clearInput();
-        await ContractApprovalBottomSheet.inputCustomAmount('2');
-
-        // Assert that custom token amount is shown
-        await Assertions.checkIfElementToHaveText(
-          ContractApprovalBottomSheet.approveTokenAmount,
-          '2',
-        );
-        // Tap next button
-        await ContractApprovalBottomSheet.tapNextButton();
-
-        // Tap approve button
-        await ContractApprovalBottomSheet.tapApproveButton();
-
+        // Approve NFT
+        await TestDApp.tapApproveERC721TokenButton();
+        await TestHelpers.delay(3000);
+        await TestDApp.tapApproveButton();
         // Navigate to the activity screen
         await TabBarComponent.tapActivity();
-        // Assert erc20 is approved
+        // Assert NFT is approved
         await Assertions.checkIfTextIsDisplayed(
           ActivitiesViewSelectorsText.CONFIRM_TEXT,
         );
