--- e2e/specs/confirmations/increase-allowance-erc20.spec.js
@@ -1,92 +1,0 @@
-'use strict';
-import { SmokeConfirmations } from '../../tags';
-import TestHelpers from '../../helpers';
-import { loginToApp } from '../../viewHelper';
-import FixtureBuilder from '../../fixtures/fixture-builder';
-import {
-  withFixtures,
-  defaultGanacheOptions,
-} from '../../fixtures/fixture-helper';
-
-import TabBarComponent from '../../pages/wallet/TabBarComponent';
-import TestDApp from '../../pages/Browser/TestDApp';
-import { SMART_CONTRACTS } from '../../../app/util/test/smart-contracts';
-import ContractApprovalBottomSheet from '../../pages/Browser/ContractApprovalBottomSheet';
-import Assertions from '../../utils/Assertions';
-import { ActivitiesViewSelectorsText } from '../../selectors/Transactions/ActivitiesView.selectors';
-import { mockEvents } from '../../api-mocking/mock-config/mock-events';
-import { buildPermissions } from '../../fixtures/utils';
-
-const HST_CONTRACT = SMART_CONTRACTS.HST;
-
-describe(SmokeConfirmations('ERC20 - Increase Allowance'), () => {
-  beforeAll(async () => {
-    if (device.getPlatform() === 'android') {
-      await TestHelpers.reverseServerPort();
-    }
-  });
-
-  it('from a dApp', async () => {
-    const testSpecificMock  = {
-      GET: [
-        mockEvents.GET.suggestedGasFeesApiGanache
-      ],
-    };
-
-    await withFixtures(
-      {
-        dapp: true,
-        fixture: new FixtureBuilder()
-          .withGanacheNetwork()
-          .withPermissionControllerConnectedToTestDapp(buildPermissions(['0x539']))
-          .build(),
-        restartDevice: true,
-        ganacheOptions: defaultGanacheOptions,
-        smartContract: HST_CONTRACT,
-        testSpecificMock,
-      },
-      async ({ contractRegistry }) => {
-        const hstAddress = await contractRegistry.getContractAddress(
-          HST_CONTRACT,
-        );
-        await loginToApp();
-        // Navigate to the browser screen
-        await TabBarComponent.tapBrowser();
-
-        await TestDApp.navigateToTestDappWithContract({
-          contractAddress: hstAddress,
-        });
-        await TestDApp.tapIncreaseAllowanceButton();
-
-        //Input custom token amount
-        await Assertions.checkIfVisible(
-          ContractApprovalBottomSheet.approveTokenAmount,
-        );
-        await ContractApprovalBottomSheet.clearInput();
-        await ContractApprovalBottomSheet.inputCustomAmount('2');
-
-        // Assert that custom token amount is shown
-        await Assertions.checkIfElementToHaveText(
-          ContractApprovalBottomSheet.approveTokenAmount,
-          '2',
-        );
-        // Tap next button
-        await ContractApprovalBottomSheet.tapNextButton();
-
-        // Tap approve button
-        await ContractApprovalBottomSheet.tapApproveButton();
-
-        // Navigate to the activity screen
-        await TabBarComponent.tapActivity();
-
-        // Assert that the ERC20 activity is an increase allowance and it is confirmed
-        await Assertions.checkIfTextIsDisplayed(
-          ActivitiesViewSelectorsText.INCREASE_ALLOWANCE_METHOD,
-        );
-        await Assertions.checkIfTextIsDisplayed(
-          ActivitiesViewSelectorsText.CONFIRM_TEXT,
-        );
-      },
-    );
-  });
-});
