+++ e2e/specs/confirmations/send-erc721.spec.ts
@@ -44,7 +44,9 @@ describe(SmokeConfirmations('ERC721 tokens'), () => {
         smartContract: NFT_CONTRACT,
         testSpecificMock,
       },
-      async ({ contractRegistry }) => {
+      // Remove any once withFixtures is typed
+      // eslint-disable-next-line @typescript-eslint/no-explicit-any
+      async ({ contractRegistry }: { contractRegistry: any }) => {
         const nftsAddress = await contractRegistry.getContractAddress(
           NFT_CONTRACT,
         );
