+++ e2e/specs/confirmations/set-approve-for-all-erc721.spec.ts
@@ -45,7 +45,9 @@ describe(SmokeConfirmations('ERC721 token'), () => {
         smartContract: NFT_CONTRACT,
         testSpecificMock,
       },
-      async ({ contractRegistry }) => {
+      // Remove any once withFixtures is typed
+      // eslint-disable-next-line @typescript-eslint/no-explicit-any
+      async ({ contractRegistry }: { contractRegistry: any }) => {
         const nftsAddress = await contractRegistry.getContractAddress(
           NFT_CONTRACT,
         );
