+++ e2e/specs/confirmations/signatures/typed-sign-v3.spec.ts
