+++ e2e/specs/confirmations/signatures/typed-sign.spec.ts
