+++ e2e/utils/Assertions.js
@@ -44,7 +44,7 @@ class Assertions {
 
   /**
    * Check if an element with the specified ID does have the specified text.
-   * @param {Promise<Detox.IndexableNativeElement>} element - The element to check.
+   * @param {Promise<Detox.IndexableNativeElement | Detox.IndexableSystemElement | Detox.NativeElement>} element - The element to check.
    * @param {string} text - The text content to check.
    * @param {number} [timeout=TIMEOUT] - Timeout in milliseconds.
    */
